requirements-cuda.txt
# face-detect requirements
opencv-python-headless~=4.5
numpy~=1.22
onnxruntime-gpu~=1.11 # for cuda11.x-cudnn8

# youtube-dl
yt-dlp~=2023.7.6 